Move ownership of LoginState to LoginViewModel
