+++ WooCommerce-Wear/src/main/java/com/woocommerce/android/wear/ui/login/FetchSiteData.kt
@@ -2,9 +2,10 @@ package com.woocommerce.android.wear.ui.login
 
 import com.woocommerce.android.wear.extensions.combineWithTimeout
 import com.woocommerce.android.wear.phone.PhoneConnectionRepository
-import com.woocommerce.android.wear.ui.login.FetchSiteData.LoginRequestState.Logged
-import com.woocommerce.android.wear.ui.login.FetchSiteData.LoginRequestState.Timeout
-import com.woocommerce.android.wear.ui.login.FetchSiteData.LoginRequestState.Waiting
+import com.woocommerce.android.wear.ui.login.LoginViewModel.LoginState
+import com.woocommerce.android.wear.ui.login.LoginViewModel.LoginState.Logged
+import com.woocommerce.android.wear.ui.login.LoginViewModel.LoginState.Timeout
+import com.woocommerce.android.wear.ui.login.LoginViewModel.LoginState.Waiting
 import com.woocommerce.commons.MessagePath.REQUEST_APP_SETTINGS
 import com.woocommerce.commons.MessagePath.REQUEST_SITE
 import kotlinx.coroutines.flow.Flow
@@ -14,7 +15,7 @@ class FetchSiteData @Inject constructor(
     private val phoneRepository: PhoneConnectionRepository,
     private val loginRepository: LoginRepository
 ) {
-    suspend operator fun invoke(): Flow<LoginRequestState> {
+    suspend operator fun invoke(): Flow<LoginState> {
         if (phoneRepository.isPhoneConnectionAvailable()) {
             phoneRepository.sendMessage(REQUEST_APP_SETTINGS)
             phoneRepository.sendMessage(REQUEST_SITE)
@@ -28,11 +29,5 @@ class FetchSiteData @Inject constructor(
                     else -> Timeout
                 }
             }
-    }
-
-    enum class LoginRequestState {
-        Logged,
-        Waiting,
-        Timeout
     }
 }
