+++ WooCommerce-Wear/src/test/java/com/woocommerce/android/wear/ui/login/LoginViewModelTest.kt
@@ -2,9 +2,9 @@ package com.woocommerce.android.wear.ui.login
 
 import androidx.lifecycle.SavedStateHandle
 import com.woocommerce.android.BaseUnitTest
-import com.woocommerce.android.wear.ui.login.FetchSiteData.LoginRequestState.Logged
-import com.woocommerce.android.wear.ui.login.FetchSiteData.LoginRequestState.Timeout
-import com.woocommerce.android.wear.ui.login.FetchSiteData.LoginRequestState.Waiting
+import com.woocommerce.android.wear.ui.login.LoginViewModel.LoginState.Logged
+import com.woocommerce.android.wear.ui.login.LoginViewModel.LoginState.Timeout
+import com.woocommerce.android.wear.ui.login.LoginViewModel.LoginState.Waiting
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.flowOf
 import org.assertj.core.api.Assertions.assertThat
@@ -21,7 +21,7 @@ class LoginViewModelTest : BaseUnitTest() {
     @Test
     fun `when login Timeouts, then view state is Timeout`() = testBlocking {
         // Given
-        var loginState: FetchSiteData.LoginRequestState? = null
+        var loginState: LoginViewModel.LoginState? = null
         whenever(fetchSiteData.invoke()).thenReturn(flowOf(Timeout))
         createSut()
         sut.viewState.observeForever { loginState = it.loginState }
@@ -34,7 +34,7 @@ class LoginViewModelTest : BaseUnitTest() {
     @Test
     fun `when login is Logged, then view state is Logged`() = testBlocking {
         // Given
-        var loginState: FetchSiteData.LoginRequestState? = null
+        var loginState: LoginViewModel.LoginState? = null
         whenever(fetchSiteData.invoke()).thenReturn(flowOf(Logged))
         createSut()
         sut.viewState.observeForever { loginState = it.loginState }
@@ -47,7 +47,7 @@ class LoginViewModelTest : BaseUnitTest() {
     @Test
     fun `when login is Waiting, then view state is Waiting`() = testBlocking {
         // Given
-        var loginState: FetchSiteData.LoginRequestState? = null
+        var loginState: LoginViewModel.LoginState? = null
         whenever(fetchSiteData.invoke()).thenReturn(flowOf(Waiting))
         createSut()
         sut.viewState.observeForever { loginState = it.loginState }
